Moved decorator and helper to utils
